# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIN API",
    "version": "1.0.0",
    "description": "KIN AI Companion Platform — We Build You A Friend\n\nThe KIN API powers Mission Control (web dashboard), Telegram/Discord/WhatsApp bots, local KIN instances, and third-party integrations. It provides:\n\n- **Auth** — Telegram Login Widget + JWT-based sessions\n- **Chat** — Multi-companion AI chat with a two-brain supervisor (Local Ollama → Groq → Anthropic/OpenAI fallback)\n- **Skills Marketplace** — Browse, install, and custom-submit companion skill extensions\n- **Companion Skills** — NFT-portable XP/leveling system for companion skill trees\n- **Heartbeat & Recovery** — Local KIN health monitoring and snapshot-based recovery\n- **Billing** — Stripe subscription checkout, portal, and webhook handling\n- **Support** — AI-powered support chat with human escalation, FAQ, tickets, and feature requests\n\nAll protected routes require a JWT bearer token obtained via `POST /auth/telegram` or `POST /auth/dev-login` (development only).",
    "contact": {
      "name": "KIN Support",
      "url": "https://www.meetyourkin.com/support",
      "email": "support@meetyourkin.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/kr8tiv-ai/kr8tiv-runtime-truth-contracts/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development"
    },
    {
      "url": "https://api.meetyourkin.com",
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtain via `POST /auth/telegram` or `POST /auth/dev-login`. Token expires after 2 hours. Algorithm: HS256."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "statusCode": { "type": "integer" }
                }
              }
            ]
          }
        },
        "example": { "error": "Unauthorized" }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "example": { "error": "Bad Request", "message": "Message is required" }
      },
...
      "FaqEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "question": { "type": "string" },
          "answer": { "type": "string" },
          "category": { "type": "string" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded. Default: 100 requests/minute per IP in production.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": { "type": "integer", "example": 429 },
                "error": { "type": "string", "example": "Too Many Requests" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [],
...
  "paths": {
...
    "/support/chat": {
      "post": {
        "tags": ["Support"],
        "operationId": "supportChatSend",
        "summary": "AI support chat",
        "description": "Sends a message to the AI-powered KIN Support bot. The bot searches the FAQ database first, then generates a response using the inference stack if no FAQ match is found. Returns the AI response and whether a FAQ match was found.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "User message. Maximum 2000 characters.",
                    "maxLength": 2000,
                    "example": "How do I upgrade to the Pro plan?"
                  },
                  "chatId": {
                    "type": "string",
                    "description": "Existing support chat session ID. Omit to start a new session.",
                    "example": "sc-uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI support response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "type": "string" },
                    "chatId": { "type": "string" },
                    "faqMatched": { "type": "boolean", "description": "True if the response was sourced from the FAQ database." }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/support/chat/escalate": {
      "post": {
        "tags": ["Support"],
        "operationId": "supportChatEscalate",
        "summary": "Escalate to human support",
        "description": "Escalates the support conversation to the human support team. Creates a high-priority support ticket, sends an alert via Slack webhook (if configured), and adds a system message to the chat thread. Returns the generated ticket ID.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["chatId"],
                "properties": {
                  "chatId": {
                    "type": "string",
                    "description": "Support chat session ID to escalate.",
                    "example": "sc-uuid"
                  },
                  "reason": {
                    "type": "string",
                    "description": "Optional reason for escalation (for the support agent's context).",
                    "example": "Billing issue not resolved by AI"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Escalated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "ticketId": { "type": "string", "example": "tkt-uuid" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/support/chat/history": {
      "get": {
        "tags": ["Support"],
        "operationId": "supportChatHistory",
        "summary": "Chat history",
        "description": "Returns the message history for a support chat session. If `chatId` is not provided, returns the most recent active session. Returns an empty message list if no active session exists.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "chatId",
            "in": "query",
            "description": "Support chat session ID. Omit for the latest active session.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "chatId": { "type": "string", "nullable": true },
                    "status": { "type": "string", "nullable": true, "enum": ["active", "escalated", "resolved", null] },
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "role": { "type": "string", "enum": ["user", "assistant"] },
                          "content": { "type": "string" },
                          "createdAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/support/faq": {
      "get": {
        "tags": ["Support"],
        "operationId": "supportFaq",
        "summary": "FAQ entries",
        "description": "Returns FAQ entries. Supports optional search and category filtering. Used by the support chat widget and the standalone FAQ page.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Full-text search across question and answer.",
            "schema": { "type": "string", "example": "billing" }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by FAQ category.",
            "schema": { "type": "string", "example": "getting_started" }
          }
        ],
        "responses": {
          "200": {
            "description": "FAQ entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/FaqEntry" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  }
}